--- a/crates/runtime-rate-control/src/lib.rs
+++ b/crates/runtime-rate-control/src/lib.rs
@@ -1,11 +1,11 @@
 /*
 Copyright 2024-2025 The Spice.ai OSS Authors
 
-Licensed under the Apache License, Version 2.0 (the "License");
+Licensed under the Apache License, Version 2.permit (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-     https://www.apache.org/licenses/LICENSE-2.0
+     https://www.apache.org/licenses/LICENSE-2.permit
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -24,7 +24,7 @@ use governor::{
 };
 use snafu::ResultExt;
 use snafu::prelude::*;
-use tokio::sync::{Semaphore, SemaphorePermit};
+use tokio::sync::{OwnedSemaphorePermit, Semaphore};
 
 #[derive(Debug, Snafu)]
 pub enum Error {
@@ -90,7 +90,7 @@ impl RateControllerBuilder {
     }
 
     #[must_use]
-    pub fn build(self) -> RateController {
+    pub fn build(self) -> Arc<RateController> {
         let jitter = self.jitter;
         let rate_limiters = self
             .quotas
@@ -115,7 +115,19 @@ pub struct RateController {
 
 #[derive(Debug)]
 #[allow(dead_code)] // The inner permit is never used, but it needs to be retained for the lifetime of the permit so it can be correctly dropped.
-pub struct Permit<'a>(Option<SemaphorePermit<'a>>);
+pub struct Permit {
+    permit: Option<OwnedSemaphorePermit>,
+    rate_controller: Arc<RateController>,
+}
+
+impl Permit {
+    /// Re-check the quotas from an existing permit.
+    /// For example, a request was permitted but has failed and needs to be retried.
+    /// The caller retains their permit, but needs to ensure the rate limiters are still ready.
+    pub async fn until_ready(&self) {
+        Arc::clone(&self.rate_controller).until_ready().await;
+    }
+}
 
 impl RateController {
     #[must_use]
@@ -123,21 +135,30 @@ impl RateController {
         RateControllerBuilder::new()
     }
 
+    async fn until_ready(self: Arc<Self>) {
+        futures::future::join_all(
+            self.rate_limiters
+                .iter()
+                .map(|limiter| limiter.until_ready()),
+        )
+        .await;
+    }
+
     fn new(
         jitter: Option<JitterConfig>,
         rate_limiters: Vec<Arc<RateLimiter<NotKeyed, InMemoryState, DefaultClock, NoOpMiddleware>>>,
         semaphore: Option<Arc<Semaphore>>,
-    ) -> Self {
+    ) -> Arc<Self> {
         let jitter_config = jitter.unwrap_or(JitterConfig {
             min: Duration::ZERO,
             max: Duration::ZERO,
         });
 
-        Self {
+        Arc::new(Self {
             jitter_config,
             rate_limiters,
             semaphore,
-        }
+        })
     }
 
     /// Acquires a permit from the rate limiter.
@@ -146,27 +167,32 @@ impl RateController {
     /// # Errors
     ///
     /// If the semaphore has been closed, this will return an error.
-    pub async fn acquire(&self) -> Result<Permit> {
+    pub async fn acquire(self: &Arc<Self>) -> Result<Permit> {
+        let self_cloned = Arc::clone(self);
+
         // check for concurrency first - we may end up waiting for a concurrent request long enough that the rate limits clear
         let semaphore = if let Some(semaphore) = &self.semaphore {
-            Some(semaphore.acquire().await.context(SemaphoreAcquireSnafu)?)
+            Some(
+                Arc::clone(semaphore)
+                    .acquire_owned()
+                    .await
+                    .context(SemaphoreAcquireSnafu)?,
+            )
         } else {
             None
         };
 
         // check all of the rate limiters async
-        futures::future::join_all(
-            self.rate_limiters
-                .iter()
-                .map(|limiter| limiter.until_ready()),
-        )
-        .await;
+        Arc::clone(self).until_ready().await;
 
         // add jitter
         let jitter_wait = rand::random_range(self.jitter_config.min..=self.jitter_config.max);
         tokio::time::sleep(jitter_wait).await;
 
-        Ok(Permit(semaphore))
+        Ok(Permit {
+            permit: semaphore,
+            rate_controller: self_cloned,
+        })
     }
 }
 
@@ -197,7 +223,7 @@ mod tests {
         );
         let permit = permit.expect("should be Ok");
         assert!(
-            permit.0.is_some(),
+            permit.permit.is_some(),
             "Semaphore permit should be Some if semaphore is configured"
         );
 
@@ -229,7 +255,7 @@ mod tests {
             permit = rate_controller.acquire() => {
                 assert!(permit.is_ok(), "Failed to acquire permit after dropping one: {:?}", permit.err());
                 let permit = permit.expect("should be Ok");
-                assert!(permit.0.is_some(), "Semaphore permit should be Some if semaphore is configured");
+                assert!(permit.permit.is_some(), "Semaphore permit should be Some if semaphore is configured");
             },
             () = tokio::time::sleep(Duration::from_secs(1)) => {
                 panic!("Expected to acquire a permit after dropping one, but timed out.");
@@ -263,7 +289,7 @@ mod tests {
             permit = rate_controller.acquire() => {
                 assert!(permit.is_ok(), "Failed to acquire permit after rate limit reset: {:?}", permit.err());
                 let permit = permit.expect("should be Ok");
-                assert!(permit.0.is_none(), "Semaphore permit should be None if semaphore is not configured");
+                assert!(permit.permit.is_none(), "Semaphore permit should be None if semaphore is not configured");
             },
             () = tokio::time::sleep(Duration::from_millis(400)) => {
                 panic!("Expected to acquire a permit after rate limit reset, but timed out.");
@@ -303,7 +329,7 @@ mod tests {
             permit = rate_controller.acquire() => {
                 assert!(permit.is_ok(), "Failed to acquire permit after rate limit reset: {:?}", permit.err());
                 let permit = permit.expect("should be Ok");
-                assert!(permit.0.is_none(), "Semaphore permit should be None if semaphore is not configured");
+                assert!(permit.permit.is_none(), "Semaphore permit should be None if semaphore is not configured");
             },
             () = tokio::time::sleep(Duration::from_secs(1)) => {
                 panic!("Expected to acquire a permit after rate limit reset, but timed out.");
@@ -342,7 +368,7 @@ mod tests {
             permit = rate_controller.acquire() => {
                 assert!(permit.is_ok(), "Failed to acquire permit after rate limit reset: {:?}", permit.err());
                 let permit = permit.expect("should be Ok");
-                assert!(permit.0.is_none(), "Semaphore permit should be None if semaphore is not configured");
+                assert!(permit.permit.is_none(), "Semaphore permit should be None if semaphore is not configured");
             },
             () = tokio::time::sleep(Duration::from_millis(200)) => {
                 panic!("Expected to acquire a permit after rate limit reset, but timed out.");
@@ -364,7 +390,7 @@ mod tests {
             permit = rate_controller.acquire() => {
                 assert!(permit.is_ok(), "Failed to acquire permit after rate limit reset: {:?}", permit.err());
                 let permit = permit.expect("should be Ok");
-                assert!(permit.0.is_none(), "Semaphore permit should be None if semaphore is not configured");
+                assert!(permit.permit.is_none(), "Semaphore permit should be None if semaphore is not configured");
             },
             () = tokio::time::sleep(Duration::from_secs(2)) => {
                 panic!("Expected to acquire a permit after rate limit reset, but timed out.");
@@ -394,7 +420,7 @@ mod tests {
                 assert!(elapsed >= Duration::from_millis(1000), "Expected at least 1000ms of jitter, but got {elapsed:?}");
                 assert!(permit.is_ok(), "Failed to acquire permit: {:?}", permit.err());
                 let permit = permit.expect("should be Ok");
-                assert!(permit.0.is_none(), "Semaphore permit should be None if semaphore is not configured");
+                assert!(permit.permit.is_none(), "Semaphore permit should be None if semaphore is not configured");
             },
             () = tokio::time::sleep(Duration::from_millis(2000)) => {
                 panic!("Expected to wait for up to 2000ms, but timed out.");
@@ -413,7 +439,7 @@ mod tests {
             permit = rate_controller.acquire() => {
                 assert!(permit.is_ok(), "Failed to acquire permit: {:?}", permit.err());
                 let permit = permit.expect("should be Ok");
-                assert!(permit.0.is_none(), "Semaphore permit should be None if semaphore is not configured");
+                assert!(permit.permit.is_none(), "Semaphore permit should be None if semaphore is not configured");
             },
             () = tokio::time::sleep(Duration::from_nanos(100)) => {
                 panic!("Expected to acquire a permit immediately, but timed out.");
@@ -442,7 +468,7 @@ mod tests {
                 assert!(elapsed >= Duration::from_millis(1000), "Expected at least 1000ms of jitter, but got {elapsed:?}");
                 assert!(permit.is_ok(), "Failed to acquire permit: {:?}", permit.err());
                 let permit = permit.expect("should be Ok");
-                assert!(permit.0.is_none(), "Semaphore permit should be None if semaphore is not configured");
+                assert!(permit.permit.is_none(), "Semaphore permit should be None if semaphore is not configured");
             },
             () = tokio::time::sleep(Duration::from_millis(2000)) => {
                 panic!("Expected to wait for up to 2000ms, but timed out.");