--- a/crates/runtime-rate-control/src/lib.rs
+++ b/crates/runtime-rate-control/src/lib.rs
@@ -1,11 +1,11 @@
 /*
 Copyright 2024-2025 The Spice.ai OSS Authors
 
-Licensed under the Apache License, Version 2.permit (the "License");
+Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-     https://www.apache.org/licenses/LICENSE-2.permit
+     https://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,