--- a/crates/runtime/Cargo.toml
+++ b/crates/runtime/Cargo.toml
@@ -129,6 +129,7 @@ runtime-datafusion-index = { path = "../runtime-datafusion-index" }
 runtime-datafusion-udfs = { path = "../runtime-datafusion-udfs" }
 runtime-object-store = { path = "../runtime-object-store" }
 runtime-table-partition = { path = "../runtime-table-partition" }
+runtime-rate-control = { path = "../runtime-rate-control" }
 rusqlite = { workspace = true, optional = true }
 rustls.workspace = true
 rustls-pemfile.workspace = true