## 📝 Summary

<!-- What does this PR change? Why is it necessary? Keep it concise. -->

* Migrates the bedrock rate limiter to using a `RateController` instead
* Adds some more logic to a `Permit` from a `RateController`, which required refactoring slightly how permits were issued with semaphore. This logic lets an existing permit re-check against the quotas while retaining their in-progress request (for example, re-checking quotas before retying a failed request).
* Fixes the concurrency for bedrock embeddings. The comment stated up to 5 requests were sent in parallel, but it looked like it was only ever 1 (sequential)? Now, it is up to the rate limiter to decide concurrency. This results in the following performance improvement:

**Before:**
```console
2025-08-15T08:04:17.955706Z  INFO runtime::accelerated_table::refresh_task: Loaded 1,000 rows (1.03 MiB) for dataset customers in 4m 21s 862ms.
```

**After:**
```console
2025-08-15T07:58:53.625147Z  INFO runtime::accelerated_table::refresh_task: Loaded 1,000 rows (1.03 MiB) for dataset customers in 20s 299ms.
```

## 🔗 Related

<!-- Link to relevant issues, discussions, or other PRs. Use "Closes #123" to auto-close issues. Omit if none. -->

* Closes #6750 
* Part of #6743